requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
discord.py==2.3.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
multidict==6.0.5
peewee==3.17.1
pi==0.1.2
smmap==5.0.1
yarl==1.9.4
